refactor: conditionally start sync update in ViewModels
This commit modifies the `startSync()` method in several ViewModels to conditionally trigger `syncManager.startSyncUpdate()` if a `loadedEquipmentType` is present in the UI state. Otherwise, it defaults to `syncManager.startSync()`.

In both scenarios, `loadedEquipmentType` is set to `null` after initiating the sync. Logging has also been added to indicate whether a regular sync or a sync update is being performed.

Affected ViewModels:
- `IPKViewModel`
- `PAAViewModel`
- `PUBTViewModel`
- `EEViewModel`
- `PTPViewModel`
- `ILPPViewModel`