fix: use correct ID for deleting BAP reports
This commit fixes a bug in the `ReportRepository` where an incorrect ID was used when deleting Berita Acara Pemeriksaan (BAP) reports.

The `deleteInspection` function in `ReportRepository.kt` now uses `moreExtraId` instead of `extraId` when calling `remoteDataSource.deleteReport` for documents of type `DocumentType.BAP`. This ensures that the correct BAP report is deleted from the remote server.