refactor: add edit mode parameter to PTP mappers
This commit updates the `toInspectionWithDetailsDomain` extension functions in `MachineMapper.kt` and `MotorDieselMapper.kt`.

An `isEdited` boolean parameter has been added to these functions. This parameter is used to set the `isEdited` field of the `InspectionDomain` object when converting the UI state to the domain model.

The `PTPViewModel.kt` has also been updated to pass the `editMode` value from `_ptpUiState` to these mapper functions.